Refactoring

+++ Auction.BusinessLogic/Services/CategoryService.cs
@@ -25,60 +25,67 @@ namespace Auction.BusinessLogic.Services
         {
             Database.Dispose();
         }
-
-        public bool IsCategoryExist(int id)
-        {
-            return Database.Categories.FindCategories(c => c.Id.Equals(id)).Any();
-        }
-
+        
         public CategoryDTO GetCategoryById(int id)
         {
-            var category = Database.Categories.GetCategoryById(id)
-                ?? throw new NotFoundException($"Category with id: {id}");
+            if(!IsCategoryExist(id))
+                throw new NotFoundException($"Category with id: {id}");
 
-            return Adapter.Adapt<CategoryDTO>(category);
+            try
+            {
+                var category = Database.Categories.GetCategoryById(id);
+                return Adapter.Adapt<CategoryDTO>(category);
+            }
+            catch (Exception)
+            {
+                throw new DatabaseException();
+            }
         }
 
         public TradingLotDTO GetLotFromCategory(int categoryId, int lotId)
         {
-            var category = GetCategoryById(categoryId);
+            try
+            {
+                CategoryDTO category = GetCategoryById(categoryId);
+                TradingLotDTO tradingLot = category.TradingLots.FirstOrDefault(lot => lot.Id == lotId)
+                    ?? throw new NotFoundException($"Lot with id: {lotId} in {category.Name}");
 
-            var lotDto = category.TradingLots.FirstOrDefault(lot => lot.Id == lotId)
-                ?? throw new NotFoundException($"Lot with id: {lotId} in {category.Name}");
-
-            return lotDto;
-        }
-
-        public IEnumerable<CategoryDTO> GetAllCategories()
-        {
-            return Adapter.Adapt<List<CategoryDTO>>(Database.Categories.FindCategories());
+                return tradingLot;
+            }
+            catch(NotFoundException ex)
+            {
+                throw ex;
+            }
+            catch (Exception)
+            {
+                throw new DatabaseException();
+            }
         }
 
         public void RemoveCategoryById(int id)
         {
             if (id == 1)
                 throw new AuctionException("You can`t delete default category");
+            if(!IsCategoryExist(id))
+                throw new NotFoundException($"Category with id: {id}");
 
-            var categoryToDelete = Database.Categories.GetCategoryById(id)
-                ?? throw new NotFoundException($"Category with id: {id}");
+            try
+            {
+                var categoryToDelete = Database.Categories.GetCategoryById(id);
 
-            if (categoryToDelete.TradingLots != null)
-                MoveLotsToCategory(categoryToDelete.TradingLots);
+                if (categoryToDelete.TradingLots.Any())
+                    MoveLotsToCategory(categoryToDelete.TradingLots);
 
-            Database.Categories.DeleteCategoryById(categoryToDelete.Id);
+                Database.Categories.DeleteCategoryById(categoryToDelete.Id);
+            }
+            catch (Exception)
+            {
+                throw new DatabaseException();
+            }
             Database.Save();
         }
 
-        private void MoveLotsToCategory(ICollection<TradingLot> tradingLots, int categoryId = 1)
-        {
-            var defaultCat = Database.Categories.GetCategoryById(categoryId);
-            tradingLots.ToList().ForEach(lot => {
-                lot.Category = defaultCat;
-                lot.CategoryId = defaultCat.Id;
-                Database.TradingLots.UpdateTradingLot(lot);
-            } );
-        }
-
+        //doesn`t work correctly remove or change
         public void EditCategory(int id, CategoryDTO category)
         {
             if (category == null)
@@ -99,35 +106,75 @@ namespace Auction.BusinessLogic.Services
             Database.Save();
         }
 
-
         public void ChangeCategoryName(int id, string name)
         {
-            var temp = Database.Categories.GetCategoryById(id)
-                ?? throw new NotFoundException($"Category with id: {id}");
+            if(!IsCategoryExist(id))
+                throw new NotFoundException($"Category with id: {id}");
             if(!IsCategoryNameFree(name))
                 throw new AuctionException($"Category {name} already exists.");
 
-            temp.Name = name;
+            try
+            {
+                var temp = Database.Categories.GetCategoryById(id);
 
-            Database.Categories.UpdateCategory(temp);
+                temp.Name = name;
+
+                Database.Categories.UpdateCategory(temp);
+            }
+            catch (Exception)
+            {
+                throw new DatabaseException();
+            }
             Database.Save();
         }
 
-        //check is category with such name is already exists and add constraint to category name as unique!!!
         public void CreateCategory(CategoryDTO category)
         {
-            if (category == null)
-                throw new ArgumentNullException(nameof(category));
             if (!IsCategoryNameFree(category.Name))
                 throw new AuctionException($"Category {category.Name} already exists.");
 
-            Database.Categories.AddCategory(new Category { Name = category.Name });
+            try
+            {
+                Database.Categories.AddCategory(new Category { Name = category.Name });
+            }
+            catch (Exception)
+            {
+                throw new DatabaseException();
+            }
             Database.Save();
         }
-        //maybe change method equeals to more specified
+
+        /// <summary>
+        /// Сhecks if there is a category with given name in the database 
+        /// </summary>
+        /// <param name="name">Name of the category</param>
+        /// <returns>True if there is no category with given name</returns>
         private bool IsCategoryNameFree(string name)
         {
-            return Database.Categories.FindCategories().FirstOrDefault(category => category.Name.Equals(name)) is null;
+            return !Database.Categories.FindCategories(category => category.Name.Equals(name)).Any();
+        }
+        /// <summary>
+        /// Сhecks if there is a category with given id in the database 
+        /// </summary>
+        /// <param name="id">Category id</param>
+        /// <returns>True if there is category with given id</returns>
+        public bool IsCategoryExist(int id)
+        {
+            return Database.Categories.FindCategories(c => c.Id.Equals(id)).Any();
+        }
+        /// <summary>
+        /// Change category of given lots in collection to category with given id
+        /// </summary>
+        /// <param name="tradingLots">Collection of lots, which category will be changed</param>
+        /// <param name="categoryId">Id of new category for lots</param>
+        private void MoveLotsToCategory(ICollection<TradingLot> tradingLots, int categoryId = 1)
+        {
+            var defaultCat = Database.Categories.GetCategoryById(categoryId);
+            tradingLots.ToList().ForEach(lot => {
+                lot.Category = defaultCat;
+                lot.CategoryId = defaultCat.Id;
+                Database.TradingLots.UpdateTradingLot(lot);
+            });
         }
     }
 }
